:recycle: refactor: separation of services for news gathering
